backend/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
